test(queues): reorganize delay tests
